Refactoring to prepare for filter panel
