+++ src/components/AnnotationDesktop/LayersPanel/ColorSettings/ColorCoding.ts
