+++ src/components/AnnotationDesktop/LayersPanel/ColorSettings/ColorPalettes.ts
