+++ src/components/AnnotationDesktop/LayersPanel/ColorSettings/colorCodings/colorByAssignment.ts
