+++ src/components/AnnotationDesktop/LayersPanel/ColorSettings/colorCodings/colorByFirstTag.ts
