+++ src/components/AnnotationDesktop/LayersPanel/ColorSettings/colorCodings/index.ts
