+++ src/components/AnnotationDesktop/LayersPanel/ColorSettings/index.ts
@@ -1,0 +1,1 @@
+export * from './ColorSettings';
