+++ src/components/AnnotationDesktop/LayersPanel/FilterSettings/index.ts
@@ -1,0 +1,1 @@
+export * from './FilterSettings';
