+++ src/i18n/de/annotation-common.json
@@ -30,6 +30,7 @@
   "newest": "Neueste",
   "oldest": "Älteste",
   "Color by": "Farbschema",
+  "Filter by": "Filter",
   "No color coding": "ohne Farbschema",
   "Public vs. Private": "Öffentlich vs. Privat",
   "Your private annotations": "Deine privaten Annotationen",
