+++ src/i18n/en/annotation-common.json
@@ -33,6 +33,7 @@
   "oldest": "oldest",
 
   "Color by": "Color by",
+  "Filter by": "Filter by",
   "No color coding": "No color coding",
   "Public vs. Private": "Public vs. Private",
   "Your private annotations": "Your private annotation",
